+++ gov.usgs.warc.iridium.sbd.decoder/src/test/java/gov/usgs/warc/iridium/sbd/decoder/parser/AllParserTests.java
@@ -1,13 +1,12 @@
 package gov.usgs.warc.iridium.sbd.decoder.parser;
 
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-import org.junit.runners.Suite.SuiteClasses;
-
 import gov.usgs.warc.iridium.sbd.decoder.parser.elements.ConfirmationTest;
 import gov.usgs.warc.iridium.sbd.decoder.parser.elements.HeaderTest;
 import gov.usgs.warc.iridium.sbd.decoder.parser.elements.LocationInformationTest;
 import gov.usgs.warc.iridium.sbd.decoder.parser.elements.PayloadTest;
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.junit.runners.Suite.SuiteClasses;
 
 /**
  * Unit test suite for running all parser tests.
@@ -17,10 +16,10 @@ import gov.usgs.warc.iridium.sbd.decoder.parser.elements.PayloadTest;
  *
  */
 @RunWith(Suite.class)
-@SuiteClasses({ BinaryParserTest.class, ConfirmationTest.class,
-		HeaderTest.class, InformationElementIdentifiersTest.class,
-		LocationDirectionTest.class, LocationInformationTest.class,
-		MessageTest.class, PayloadTest.class, SessionStatusTest.class,
+@SuiteClasses({ ConfirmationTest.class, HeaderTest.class,
+		InformationElementIdentifiersTest.class, LocationDirectionTest.class,
+		LocationInformationTest.class, MessageTest.class, PayloadTest.class,
+		SbdParserTest.class, SessionStatusTest.class,
 		StationDataTypesTest.class })
 public class AllParserTests
 {