+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  * Update to Guava 27.1
  * Update to JDT Annotation 2.2.200
  * Update to Commons-CSV 1.6
+ * Rename BinaryParser to SbdParser
  
 ## 1.1.0 - 11/16/2018
  * Update .travis.yml to build OpenJDK11 and all branches